Feature: Made all int variables into float

+++ src/CalculatorTests.py
@@ -17,38 +17,38 @@ class MyTestCase(unittest.TestCase):
     def test_add_method_calculator(self):
         test_add = CsvReader('/src/Unit Test Addition.csv').data
         for row in test_add:
-            self.assertEqual(self.calculator.add(row['Value 1'], row['Value 2']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.add(row['Value 1'], row['Value 2']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
 
     def test_subtract_method_calculator(self):
         test_sub = CsvReader('/src/Unit Test Subtraction.csv').data
         for row in test_sub:
-            self.assertEqual(self.calculator.subtract(row['Value 1'], row['Value 2']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.subtract(row['Value 1'], row['Value 2']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
 
     def test_multiplication_method_calculator(self):
         test_multiplication = CsvReader('/src/Unit Test Multiplication.csv').data
         for row in test_multiplication:
-            self.assertEqual(self.calculator.multiplication(row['Value 1'], row['Value 2']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.multiplication(row['Value 1'], row['Value 2']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
 
     def test_division_method_calculator(self):
         test_division = CsvReader('/src/Unit Test Division.csv').data
         for row in test_division:
-            self.assertEqual(self.calculator.division(row['Value 1'], row['Value 2']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.division(row['Value 1'], row['Value 2']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
 
     def test_square_method_calculator(self):
         test_square = CsvReader('/src/Unit Test Square.csv').data
         for row in test_square:
-            self.assertEqual(self.calculator.square(row['Value 1']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.square(row['Value 1']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
 
     def test_root_method_calculator(self):
         test_root = CsvReader('/src/Unit Test Square Root.csv').data
         for row in test_root:
-            self.assertEqual(self.calculator.root(row['Value 1']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.root(row['Value 1']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
 
 
 if __name__ == '__main__':
